pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "boids"
version = "0.1.0"
description = "A package for displaying boids simulations"
authors = ["The Alan Turing Institute <hut23@turing.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.5.2"
numpy = "^1.23.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.961"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
types-PyYAML = "^6.0.10"

[tool.poetry.scripts]
boids = "boids.view:main"

[tool.isort]
profile = "black"
